All finish

+++ exercises/23_conversions/using_as.rs
@@ -10,11 +10,10 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    // Conversion explicite de la longueur du tableau en f64 pour éviter une division entière
+    total / values.len() as f64
 }
 
 fn main() {
@@ -31,3 +30,4 @@ mod tests {
         assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
     }
 }
+
